Add the ability to run the check function against a mongo document field in the plugin
We add a `reject` property to the options object in the plugin api.

Currently there are two functions added, one attempts to add a validation function to a schema field declaration and the other uses a pre validation hook. If anyone has a better understanding of mongoose than I, it would be great to shoot the breeze on this. Moongoose docs suck.